// src/interfaces/marathon.ts
export interface IMarathonDetailData {
  user: {
    email: string;
    firstName: string;
    lastName: string;
    mobileNumber: string;
    gender: string;
    marathonSkill: string;
  };
  contactInformation: {
    email: string;
    firstName: string;
    lastName: string;
    mobileNumber: string;
  };
  additionalInformation: string;
}